python-requirements.txt
# Updated requirements for the email aggregator with RAG capabilities
imapclient==2.3.1
mail-parser==3.16.0
elasticsearch==8.10.0
chromadb==0.3.26
sentence-transformers==2.2.2
discord.py==2.3.2
streamlit==1.26.0
tf-keras
torch
transformers==4.33.3
numpy==1.24.4
accelerate==0.22.0
fastapi==0.99.1
uvicorn==0.23.2
pydantic==1.10.9
python-dotenv==1.0.0
groq
